+++ packages/plugins/communication-explorer/src/headless/services/auto-doc-filtering/index.ts
@@ -1,0 +1,2 @@
+export { applyConnectionFilters } from './apply-autodoc-filters'
+export { filterIEDsByBays } from './filters'